Backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==3.0.3
PyJWT==2.9.0
psycopg2-binary==2.9.10
google-auth==2.36.0
